Move ShapeEnv config out of dynamo
Previously there was a circular dependency between fx and dynamo that happened
to work out since ShapeEnv didn't access the config at module init time.

[ghstack-poisoned]
