+++ test/dynamo/test_dynamic_shapes.py
@@ -3,6 +3,7 @@ import unittest
 import warnings
 
 from torch._dynamo import config
+from torch.fx.experimental import _config as fx_config
 from torch._dynamo.testing import make_test_cls_with_patches
 from torch.testing._internal.common_utils import TEST_Z3
 
@@ -44,8 +45,8 @@ def make_dynamic_cls(cls):
         suffix,
         (config, "assume_static_by_default", False),
         (config, "specialize_int", False),
-        (config, "translation_validation", TEST_Z3),
-        (config, "check_shape_env_recorded_events", True),
+        (fx_config, "translation_validation", TEST_Z3),
+        (fx_config, "check_shape_env_recorded_events", True),
         xfail_prop="_expected_failure_dynamic",
     )
