+++ test/dynamo/test_exc.py
@@ -192,11 +192,13 @@ ReluCompileError:""",
 
     @skipIf(not TEST_Z3, "z3 not installed")
     @torch._dynamo.config.patch(
-        inject_EVALUATE_EXPR_flip_equality_TESTING_ONLY=True,
         assume_static_by_default=False,
+        suppress_errors=False,
+    )
+    @torch.fx.experimental._config.patch(
+        inject_EVALUATE_EXPR_flip_equality_TESTING_ONLY=True,
         translation_validation=True,
         translation_validation_no_bisect=True,
-        suppress_errors=False,
     )
     def test_trigger_on_error(self):
         from torch.fx.experimental.validator import ValidationException
@@ -262,10 +264,12 @@ Failed Source Expressions:
 
     @skipIf(not TEST_Z3, "z3 not installed")
     @torch._dynamo.config.patch(
-        inject_EVALUATE_EXPR_flip_equality_TESTING_ONLY=True,
         assume_static_by_default=False,
-        translation_validation=True,
         suppress_errors=False,
+    )
+    @torch.fx.experimental._config.patch(
+        inject_EVALUATE_EXPR_flip_equality_TESTING_ONLY=True,
+        translation_validation=True,
     )
     def test_trigger_bisect_on_error(self):
         from torch.fx.experimental.validator import BisectValidationException
