+++ torch/_dynamo/config.py
@@ -141,9 +141,6 @@ replay_record_enabled = os.environ.get("TORCH_COMPILE_DEBUG", "0") == "1"
 # Rewrite assert statement in python with torch._assert
 rewrite_assert_with_torch_assert = True
 
-# [@compile_ignored: debug] Show a warning for every specialization
-print_specializations = False
-
 # Disable dynamo
 disable = os.environ.get("TORCH_COMPILE_DISABLE", False)
 
@@ -277,30 +274,6 @@ report_all_guard_failures = False
 # [@compile_ignored: debug] root folder of the project
 base_dir = dirname(dirname(dirname(abspath(__file__))))
 
-# [@compile_ignored: debug] Uses z3 for validating the guard optimizations transformations.
-translation_validation = (
-    os.environ.get("TORCHDYNAMO_TRANSLATION_VALIDATION", "0") == "1"
-)
-# Timeout (in milliseconds) for z3 finding a solution.
-# [@compile_ignored: debug]
-translation_validation_timeout = int(
-    os.environ.get("TORCHDYNAMO_TRANSLATION_VALIDATION_TIMEOUT", "600000")
-)
-# Disables bisection for translation validation.
-#
-# Translation validation bisection is enabled by default, if translation validation
-# is also enabled. This should help finding guard simplification issues. However,
-# since validation uses Z3 for bisecting, it might take a lot of time.
-#
-# Set this configuration option so as to avoid bisecting.
-# [@compile_ignored: debug]
-translation_validation_no_bisect = (
-    os.environ.get("TORCHDYNAMO_TRANSLATION_NO_BISECT", "0") == "1"
-)
-# Checks whether replaying ShapeEnv events on a freshly constructed one yields
-# the a ShapeEnv with the same state. This should be used only in testing.
-check_shape_env_recorded_events = False
-
 # Trace through NumPy or graphbreak
 trace_numpy = True
 
@@ -361,11 +334,6 @@ capture_func_transforms = True
 # used for testing
 inject_BUILD_SET_unimplemented_TESTING_ONLY = False
 
-# wraps (un)equalities with 'Not' class after recording the correct expression
-# in the FX graph. This should incorrectly construct the divisible and replacement
-# lists, and incorrectly issue guards.
-inject_EVALUATE_EXPR_flip_equality_TESTING_ONLY = False
-
 _autograd_backward_strict_mode_banned_ops = [
     "stride",
     "requires_grad",
@@ -383,6 +351,6 @@ _autograd_backward_strict_mode_banned_ops.extend(
 # WARNING: this is an experimental flag and is subject to change.
 _experimental_support_context_fn_in_torch_utils_checkpoint = False
 
-from .config_utils import install_config_module
+from torch.utils._config_module import install_config_module
 
 install_config_module(sys.modules[__name__])
