+++ torch/utils/_config_module.py
